api/requirements.txt
Flask>=2.0
python-dotenv>=0.15
Flask-CORS>=3.0.0